# docs/codex/rituvia-production-2026-07-23/05_API_CONTRACTS.openapi.yaml
openapi: 3.1.0
info:
  title: RITUVIA Production API
  version: 1.0.0-draft
  description: Server-authoritative API contract. Adapt paths to the existing repository without weakening invariants.
servers:
  - url: https://staging.example.invalid/api
security:
  - cookieSession: []
...
paths:
  /v1/auth/email/start:
    post:
      tags: [Auth]
      security: []
      operationId: startEmailSignIn
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, locale]
              properties:
                email: {type: string, format: email, maxLength: 254}
                locale: {type: string, enum: [en, zh-CN]}
                returnPath: {type: string, maxLength: 300}
      responses:
        '202': {description: Uniform accepted response; never reveals account existence}
        '429': {$ref: '#/components/responses/RateLimited'}
...
  /v1/numerology/calculate:
    post:
      tags: [Readings]
      security: []
      operationId: calculateNumerology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [birthDate]
              properties:
                birthDate: {type: string, format: date}
                latinName: {type: string, maxLength: 120}
      responses:
        '200': {description: Deterministic results and visible calculation steps}
...
  /v1/rituals/start:
    post:
      tags: [Sanctuary, Credits]
      operationId: startRitual
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [objectCode]
              properties:
                objectCode: {type: string}
                intentionId: {type: string, format: uuid}
      responses:
        '201': {description: Ritual session created; consumable pass atomically consumed/reserved}
        '403': {description: No entitlement or ritual pass}
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: __Host-rituvia_session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 255}
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 512}
  responses:
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: {schema: {type: integer}}
...